SDIT-1803 Create Activities specific NOMIS client
